[Feat] #19 - SessionListViewController View 코드 CustomListTableViewCell로 이전
