+++ Carmunication/Models/DummyGroup/DummyGroup.swift
@@ -1,0 +1,18 @@
+//
+//  DummyGroup.swift
+//  Carmunication
+//
+//  Created by 김동현 on 2023/10/03.
+//
+
+import Foundation
+
+struct DummyGroup {
+    var groupTitle: String = "(주)좋좋소"
+    var subTitle: String = "늦으면 큰일이 나요"
+    var isDriver: Bool = false
+    var startPoint = "출발지"
+    var endPoint = "도착지"
+    var startTime = "08:30"
+    var crewCount = 4
+}
